requirements.txt
fastapi>=0.116.0
uvicorn>=0.35.0
pydantic>=2.11.0
pyyaml>=6.0.2
pytest>=8.3.0
